task12

+++ rosalind/src/main/scala/tasks/Task12.scala
@@ -7,10 +7,10 @@ object Task12 extends Solver {
   import Solver._
 
   //Convert a DNA string to a number.
-  def hash(seq: String): Int = {
+  def hash(seq: String): Long = {
     val subtMap = Map('A' -> 0, 'C' -> 1, 'G' -> 2, 'T' -> 3)
-    seq.reverse.foldLeft((0,0)){case ((acc,expon),c) =>
-      (acc+Math.pow(4,expon).toInt*subtMap(c),expon+1)
+    seq.reverse.foldLeft((0L,0)){case ((acc,expon),c) =>
+      (acc+Math.pow(4,expon).toLong*subtMap(c),expon+1)
     }._1
   }
